refactor: improvements on the toggle of Google Font Local Host
