--- assets/apps/customizer-controls/src/builder-local-google-fonts/Control.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import ToggleComponent from '../toggle/ToggleComponent';
-
-const Control = ({ control }) => <ToggleComponent control={control} />;
-
-export default Control;
